requirements.txt
flask
flask-sqlalchemy
gunicorn
email-validator
numpy
pandas
plotly
netcdf4
psycopg2-binary
pyjulia
julia
email-validator
flask
flask-login
flask-migrate
flask-sqlalchemy
flask-wtf
numpy
pandas
plotly
psycopg2-binary
python-dotenv
pytz
requests
email-validator
flask
flask-login
flask-migrate
flask-sqlalchemy
flask-wtf
python-dotenv
openpyxl
pandas
xlsxwriter
